Prefer Jetson v4l2m2m encoder
