+++ ffmpeg_utils.py
@@ -4,6 +4,39 @@ import threading
 from typing import List, Optional, Tuple
 
 
+def detect_encoder() -> str:
+    """Detect and return a usable H.264 encoder.
+
+    Preference order:
+    1. ``h264_v4l2m2m`` (Jetson hardware encoder)
+    2. ``h264_nvmpi``
+    3. ``libx264``
+
+    ``h264_omx`` is intentionally skipped due to reliability issues. A
+    ``RuntimeError`` is raised if no suitable encoder is found.
+    """
+
+    try:
+        result = subprocess.run(
+            ["ffmpeg", "-encoders"],
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+            text=True,
+        )
+        encoders = result.stdout
+        if "h264_v4l2m2m" in encoders:
+            return "h264_v4l2m2m"
+        if "h264_nvmpi" in encoders:
+            return "h264_nvmpi"
+        if "libx264" in encoders:
+            return "libx264"
+    except Exception:
+        pass
+    raise RuntimeError(
+        "❌ No usable H.264 encoder found (looked for h264_v4l2m2m, h264_nvmpi, libx264)."
+    )
+
+
 def run_ffmpeg_command(cmd: List[str], timeout: int = 15) -> Tuple[int, str, str]:
     """Run an FFmpeg command with realtime stderr streaming.
